Move overlap tests into a separate test file
